+++ SB03-Findex-Team3/src/main/java/com/sprint/findex/service/AutoSyncConfigService.java
@@ -3,7 +3,7 @@ package com.sprint.findex.service;
 import com.sprint.findex.dto.request.AutoSyncConfigUpdateRequest;
 import com.sprint.findex.dto.request.AutoSyncQueryParams;
 import com.sprint.findex.dto.response.AutoSyncConfigDto;
-import com.sprint.findex.dto.response.CursorPageResponseAutoSyncConfigDto;
+import com.sprint.findex.dto.response.cursor.CursorPageResponseAutoSyncConfigDto;
 
 public interface AutoSyncConfigService {
