+++ SB03-Findex-Team3/src/main/java/com/sprint/findex/service/IndexDataService.java
@@ -6,7 +6,7 @@ import com.sprint.findex.dto.dashboard.RankedIndexPerformanceDto;
 import com.sprint.findex.dto.request.IndexDataCreateRequest;
 import com.sprint.findex.dto.request.IndexDataQueryParams;
 import com.sprint.findex.dto.request.IndexDataUpdateRequest;
-import com.sprint.findex.dto.response.CursorPageResponseIndexData;
+import com.sprint.findex.dto.response.cursor.CursorPageResponseIndexData;
 import com.sprint.findex.dto.response.IndexDataDto;
 import com.sprint.findex.entity.Period;
 import java.util.List;
