+++ SB03-Findex-Team3/src/main/java/com/sprint/findex/service/SyncJobService.java
@@ -2,7 +2,7 @@ package com.sprint.findex.service;
 
 import com.sprint.findex.dto.request.IndexDataSyncRequest;
 import com.sprint.findex.dto.request.SyncJobQueryParams;
-import com.sprint.findex.dto.response.CursorPageResponseSyncJobDto;
+import com.sprint.findex.dto.response.cursor.CursorPageResponseSyncJobDto;
 import com.sprint.findex.dto.response.SyncJobDto;
 import java.util.List;
 import reactor.core.publisher.Mono;
